results
some examples of standard data output for a small portion of integers beginning with 1. these were run on my home lab which is a WIP (ubuntu server VM on proxmox hypervisor, running on an old laptop) so the speeds do no reflect the potential speed of the script, my benchmark for these calculations within this range is around 3-4 seconds on my macbook.